Convert HTML model to ES6 Classes
I caved on this one. Generally speaking I am not a proponent of using
classes because you should not write JavaScript like it's Java, and
functions offer a far more flexible and far less verbose way of
accomplishing many of the same things.

However, in this case, what I need to do is group the storage of some
data (a direction step, a direction leg) with its presentation. I guess
that I could probably create an object with a function and override the
toString prototype, but that's essentially what the ES6 syntax does and
its far less intuitive. I think that in this case not using classes
would be dogmatic and silly.

Shame I have to use the "new" keyword in maps now. Oh well.