Add fieldset for trip start and end
These are the dropdowns that determine where the trips start and end.
They will soon be followed by the ability to specify which trips are
getting on and off at each location, so that the backend can generate a
packet with appropriate "who's on the bus" numbers for each stop.