Enable multiple-stop packets
This is the simplest useful version of the product! You can choose from
any number of available stops, and create a packet that provides the
directions to take you to each stop in order. At each stop, the packet
will contain the special instructions to the driver to reach that spot.

I still need to do trippees on and off at each stop, but this is enough
to start getting feedback. It also still needs some QOL improvements,
like viewing each set of directions individually to verify them.