Add preliminary button to update trip on/off selectors
When the function is triggered, it updates the two associated selectors
for each trip that let you pick where the trip will get on and off. If
the option that was previously selected is no longer present, it will
default to the first or last option in the list, depending on which
selector it is.

Right now these are triggered on a button, but there will probably be a
cleverer hook soon.