Make it possible to edit trip boardings
This is a deceptively complicated change. The main trick is that the
function that turns queries into form values is run twice; because the
second half of the form depends on your selections in the first half,
the function to update your choices in the second half has to run
*after* those choices have been populated but before selections in the
second half of the form are made.

A slightly different form setup might avoid this, which I will consider
at a later period. I also have to figure out why vim formats JS inside
HTML so weirdly.