Unify packet sections into one list
In a mild deviation from the sample packet I had each leg of the trip
grouped into sections. It turns out to be conceptually simpler to just
render both the destinations and the trip directions as elements in the
same ordered list. Now the destinations contain all the info about trips
getting on and off, as well as special instructions.