Add a button to regenerate the packets
To accomplish this, I added a /regenerate route that triggers the
regenerations. I think this action should probably move to a worker
thread or a child process at some point so that there is no risk of it
blocking the event loop, but this should be more than sufficient for the
time being.

The error handling is currently a little subpar here, because this
regeneration option could occur on a packet that has had relevant trips
or destinations deleted. Deleted trips would be fine (it says the trip
name and "undefined" for number of people) but stops would be weird.
Need a more graceful way of handling that.