Display which packets errored out upon regeneration
This is a really important addition for making this feature usable. It's
a far better UX to have the backend explain which packets are
problematic so that you can fix them, rather than just saying: looks
broken.

Not all implausible that they could add delete a stop or a trip and then
try to regenerate the packets.