Fix NPE that occurs if stop changes coordinates
If someone changed the coordinates of a stop, the directions list will
still have directions that start or end at coordinates that are no
longer associated with a stop. The SQL query should simply filter those
out, since they are no longer relevant (and can be safely deleted as
well, might be worth having a trigger do that).