Fix packet regeneration error handling
Previously it was communicating the wrong names of the packets that
failed if the promises did not all return in the same order that they
were called (i.e. if some of them made network calls and others did
not). Fixed this by changing the algorithm to filter out the successful
promises after having already determined the names of the failed
promises.