Setup application for fly.io deployment
Using a new managed deployment system this year because I think it's
warranted, especially since fly.io supports SQLite. This was a bit of a
pain but ultimately it's a just a Dockerfile and a little fly config
that says "here's the persistent volume."

I will write up the documentation about its deployment soon.